refactor!: remove `from_(yaml|ron)_(str|bytes)` functions
